Move WebExtension::Helper into separate file
