+++ .rubocop_todo.yml
@@ -1,6 +1,6 @@
 # This configuration was generated by
 # `rubocop --auto-gen-config`
-# on 2019-12-04 20:56:11 +0100 using RuboCop version 0.77.0.
+# on 2019-12-06 21:33:12 +0100 using RuboCop version 0.77.0.
 # The point is for the user to remove these configuration records
 # one by one as the offenses are removed from the code base.
 # Note that changes in the inspected code, or installation of new
@@ -10,6 +10,11 @@
 # Configuration parameters: Severity.
 Metrics/AbcSize:
   Max: 22
+
+# Offense count: 3
+# Configuration parameters: CountComments, ExcludedMethods, Severity.
+Metrics/MethodLength:
+  Max: 19
 
 # Offense count: 5
 # Configuration parameters: .
