+++ lib/sidekiq/undertaker.rb
@@ -1,11 +1,7 @@
 # frozen_string_literal: true
 
 require "sidekiq/undertaker/version"
-require "sidekiq/undertaker/dead_job"
-require "sidekiq/undertaker/job_filter"
 require "sidekiq/undertaker/web_extension"
-require "sidekiq/undertaker/bucket"
-require "sidekiq/undertaker/job_aggregator"
 
 begin
   require "sidekiq/web"
